requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
pandas>=1.3.0
psycopg2-binary>=2.8.0
mysql-connector-python>=8.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
Werkzeug>=2.0.0